:art: move out-of-band model definitions to own module
